Remove GPU tests analyze exclusion
Removes //content/test/gpu/ from the file exclusions of the default
analyze config. It is suspected that this was originally added to ensure
that changes to the GPU Telemetry-based test harness or the associated
expectation files would cause the relevant tests to run. However, this
is already handled properly by the analyze step, and the exclusion
causes unrelated trybots to compile, which wastes time and resources.

Bug: 370968805
Change-Id: I9497b1a6792e526cc3d23aebb4b31c943fe5e838
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5922746
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1367640}
NOKEYCHECK=True
GitOrigin-RevId: 8776a66654221f3822e117fe71a84222901048d5