Specify all features in the MemoryPurgeInBackground testing config.
With this CL, each group of the MemoryPurgeInBackground study sets the
state of all features explicitly, even if the state is the same as in
the code. I learned that it is best to have studies that set all
relevant features explicitly, to avoid a change of group behavior
if the default state in the code changes (and the testing config must
match the server-side config exactly).

This is a follow-up to crrev.com/c/5938697

Bug: 40287434
Change-Id: Ie518614eb5557a491ec04b682cf54ae812bd1479
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5941919
Auto-Submit: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Francois Pierre Doray <fdoray@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1371657}
NOKEYCHECK=True
GitOrigin-RevId: 4de711509aab9b23acbae82a58aecc4af969ffb6