Add flag to skip pagehide-in-commit when navigating to DSE
When introducing proactive BI swap for BFCache, we also introduced pagehide-in-commit, to trigger pagehide & visibilitychange events before the new document commits despite doing a cross-BrowsingInstance navigation, which would typically run those post-commit. This is to prevent web-breaking changes for pages that expect the order of events to stay the same (new document commits after pagehide runs).

Recently we discovered that this is slowing down navigations to DSE, and they probably don't depend on the order above. The plan is to experiment skipping pagehide-in-commit for navigations to DSE to see if it improves navigation latency.

Bug: 375385416
Change-Id: I9d1c3dee470ebb08b0a7060b5e5712483ee35c90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5960961
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1373656}
NOKEYCHECK=True
GitOrigin-RevId: 35d978b6991c5ca367415c3bbf086661f77d621c