Enable ffmpeg H.264 decoding by default on Android
Experiment results are strongly positive, so launch! Note: This
is still only compiled on non-ARM devices.

This also fixes a test that was not running with the field trial
config and needed updating. The storage type of the frames was
incorrect since they are wrapped by FrameBufferPool.

Bug: 348538294
Change-Id: If995921875aef04e6641455639c87586747724ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5973276
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1375625}
NOKEYCHECK=True
GitOrigin-RevId: cf08c495da05a3e9477f1a4c26a91b0f85a076fc