chromeos: Enable feature BirchWeather by default
This is a precondition to marking the Finch config "launched". Add one
check for a null user manager, as there are some tests that enter
overview mode and try to trigger a weather fetch, but do not have a
user manager instance.

Bug: 358132644
Test: ash_unittests
Change-Id: I045334216e12476e8eb5d3f511266836671e704f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5980408
Reviewed-by: Toni Barzic <tbarzic@google.com>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1375981}
NOKEYCHECK=True
GitOrigin-RevId: a6985a78ef072f23f56c232b1f49155383c85b2e