Implement "PrefetchManagerUseNetworkContextPrefetch" feature
Add a feature to use the new NetworkContext::Prefetch() method from
PrefetchManager instead of the current implementation. This will enable
gathering metrics on how effectively the predicted requests are matching
the eventual real requests.

For this to do anything, both "PrefetchManagerUseNetworkContextPrefetch"
and "NetworkContextPrefetch" features to be enabled. On desktop
platforms "LoadingPredictorUseOptimizationGuide" and
"LoadingPredictorPrefetch" also need enabling (they are on by default on
Android).

The new code path completely replaces the old one when enabled, so the
feedback that is normally available from PrefetchManager is not
available. A new function, PerformNetworkContextPrefetch() calculates
the expected value of the network::ResourceRequest that the render
process will use and dispatches it to the NetworkContext.

The requirements for matching headers are much more stringent with the
new code path, so move reduce_accept_language_utils.* from
//content/browser to //content/public/browser so that the proper value
of the `Accept-Language` header can be calculated.

Add the new feature to fieldtrial_testing_config.json in preparation for
running an experiment.

Example of the difference between the HTTP headers sent with and
without the new feature enabled:

 accept-encoding: gzip, deflate, br, zstd
-accept-language: en-GB,en-US;q=0.9,en;q=0.8
+accept-language: en-GB,en;q=0.9
+accept: text/css,*/*;q=0.1
 priority: u=4, i
 purpose: prefetch
+referer: https://www.amazon.co.jp/
+sec-ch-ua-mobile: ?0
+sec-ch-ua-platform: "Linux"
+sec-ch-ua: "Not A(Brand";v="8", "Chromium";v="132"
 sec-fetch-dest: style
 sec-fetch-mode: no-cors
 sec-fetch-site: cross-site

Headers marked with "+" are different or new with the new feature.

BYPASS_LARGE_CHANGE_WARNING=Cannot reasonably be broken up

Bug: 342445996
Change-Id: Ice4b43b6594bf18edcd77471c82cbc287ba7a449
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5974809
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1376223}
NOKEYCHECK=True
GitOrigin-RevId: c98829dcfc332b2439aa6978785895d33e0b1ac8