Add Regmon fieldtrial testing config
Add Regmon fieldtrial testing config, as this is a prerequisite to feature rollout.

Also includes a minor fix to the feature's browser test, as the testing config change causes the test to fail since the feature is no longer disabled-by-default in tests.

Bug: b:376122549
Change-Id: Id8e4486542c0da2777f5ea05baa2fdfa1b532bce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5979232
Reviewed-by: Chris Mullins <crmullins@chromium.org>
Commit-Queue: Aiden Chiavatti <chiav@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1381508}
NOKEYCHECK=True
GitOrigin-RevId: 13952ec9fc7d42fc53e34a8ae11bd8f36c84de32