TabDragging tests: Enable *DragToSeparateWindow* tests for ozone/wayland
We need to rewrite the tests that rely on DragTabAndNotify() /
QuitDraggingObserver.Wait() returning immediately when attaching to a
new context, even when the tab dragging session continues running, as
that doesn't work with fallback tab dragging. Instead, for these tests
we introduce BrowserRemovedWaiter, which they can use to wait for the
browser containing the dragged tabs being removed when being attached to
another browser, and then run the code that leads to the end of the drag
session.

Bug: 40238145
Change-Id: I6f1d9cf1970dec729d702e98ff5bd7040f0cb92a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5850730
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Commit-Queue: Max Ihlenfeldt <max@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1382254}
NOKEYCHECK=True
GitOrigin-RevId: c48aee37356cd16272b9a19be1f8d92545307260