Rework locking controls from navigations.
Page load and navigation events are not guaranteed to be delivered for one navigation at a time. Instead they can be interleaved. This is causing a problem when a slow navigation starts before the previous navigation is completed. After the 3 second wait, we erroneously unlocked the browser controls when they should not be.

To fix this, all page load events were switched to navigation events. And we now track navigation ids in a set, and only unlock when the set is empty. Special casing for same document navigations was added to handle the difference between navigation events and page load events. This approach has the potential danger of leaving browser controls locked forever if we miss an event.

To mitigate this risk, this CL adds a kill switch and a histogram that counts the number of outstanding navigations when a navigation completes.

Bug: 379652406
Change-Id: I6a837b7f7b2103c6f811e130bf582ac38c66c763
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6040200
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1388487}
NOKEYCHECK=True
GitOrigin-RevId: 2a80258ed798263eeff7ca49f3c13d389eb87df8