[chromeperf] Reduce the swarming task hard timeout to 4 hours
The current timeout is the same as the luci build timeout. If any swarming task times out, the luci build will not have time to finish the upload steps.

This CL reduces the timeout for swarming tasks. If some of the swarming tasks time out, the luci build will have time to proceed and upload the perf results for those healthy shards.

Ignore-Freeze: fix a builder which is failing at a 40% rate.
Bug: b/380929167
Change-Id: I7ae943ddff08382d86b134a0e31a1ac95c11f347
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6052617
Reviewed-by: John Chen <johnchen@chromium.org>
Commit-Queue: Wenbin Zhang <wenbinzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#1388540}
NOKEYCHECK=True
GitOrigin-RevId: a5a57a4d70be69af1223e5a5b867fb3179fb3893