[etw tracing] Use the ETW tracing service in Chrome
The browser will attempt to connect to the system ETW tracing service if
it is installed and the WindowsSystemTracing feature is enabled. Traces
will therefore include system-wide events sourced from ETW.

To make this possible, OnTracingService{Started,Stopped} are added to
ContentBrowserClient so that the embedder may take action on these
events.

Bug: 41497361
Change-Id: I3c68e426f241fa00b7b769cbd6775e8714246c22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6038066
Commit-Queue: Greg Thompson <grt@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1388962}
NOKEYCHECK=True
GitOrigin-RevId: 59bf58e111e71b390c03a83c3a49fa93b633e0cd