windows startup: OVERRIDE_PREFETCH_PARAMETER on acquiring singleton
With this change, run a field trial to determine the impact of
OVERRIDE_PREFETCH_PARAMETER on the acquiry of the singleton.

This should effectively solve the case where short-lived browser
processes are polluting the App Launch Prefetch history. Local tracing
leads us to believe that recent versions of Windows implicitly support
this but we decided to experiment with this newly discovered API
nonetheless to verify how this fares in the wild.

Bug: 325307453, 380088804
Change-Id: I515d983c6f558e4e896aadf4796b48487236f5b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5939457
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Commit-Queue: Sean Maher <spvw@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1389076}
NOKEYCHECK=True
GitOrigin-RevId: 5693c9daaeb9ad34a00575195caf7f8a16342c35