Add field trial config for event based logs
EventBasedLogUpload feature is going to be launched with a Finch field
trial. Add the field trial config according to the instructions in
go/finch-best-practices.

We don't need to set the LogUploadEnabled policy in the browser test
anymore. This config enables the feature on browser tests.

Bug: b:381064170
Change-Id: I6053d909a3bffebdcbe57c6261f4e0c196e58f4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6054282
Commit-Queue: Irem Uguz <iremuguz@google.com>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1389814}
NOKEYCHECK=True
GitOrigin-RevId: 64542391b07db49f8a9d05f31426a00b9cba72f1