Experiment with disabling async navigation interception
Launched in 2019, the AsyncNavigationIntercept feature causes ExternalNavigationHandler to evaluate whether the navigation should instead launch an app to run in parallel to the navigation.

This optimization is incorrect, because if the redirect arrives before the initial navigation is evaluated, we just evaluate the redirect URL twice and forget about the initial URL. There's no trivial way to fix this as the NavigationHandle is mutated throughout the navigation and you can't copy a NavigationHandle.

We should run an experiment to see how much this optimization actually helps in practice these days, and consider unlaunching it for correctness and complexity reasons.

Bug: 381535042
Change-Id: Ida86546851cc7bd7131130027820b95ad8db4450
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6055004
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1390460}
NOKEYCHECK=True
GitOrigin-RevId: b236176cc0997e9f483c3d28a8ff335f920f730d