[iOS]Introduce interruptible-coordinator-stop-synchronously
This CL introduces the feature flag
"interruptible-coordinator-stop-synchronously".

While this flag is disabled, there is no change in the execution
of the code.

When this flag is enabled, the add account signin manager
don’t send any completion to `identityInteractionManager`. Instead
it executes the completion synchronously, after having called sent
the request to stop the add account view.

This CL also updates the fake so that it checks that the received
completion is nil, as is now expect no value.

Once the flag is removed, we can change the API to remove
the completion block parameter entirely.

Bug: 381444097
Change-Id: I1fc9aebadc5dcb572df0478eb7620fae5f7bbf54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6023495
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1392842}
NOKEYCHECK=True
GitOrigin-RevId: 35ee54f3e6010d99cec826b01819cbe3ee958a68