Binary-launch HistoryEmbeddings/Answers keeping feature override bypass
This CL default-enables the LaunchedHistoryEmbeddings and
LaunchedHistoryEmbeddingsAnswers features, effectively client-side
launching history embeddings with answers for desktop platforms in US.
If the HistoryEmbeddings or HistoryEmbeddingsAnswers features are
overridden, the launch logic is bypassed.

The ChromeOS feature management and HistorySearchSettingsVisibility are
not affected by this change.

Bug: 381108241
Change-Id: I848b33a7fac2b8e3760c63c214b4868b072ea86a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6077916
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Orin Jaworski <orinj@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1396033}
NOKEYCHECK=True
GitOrigin-RevId: f20a2099a1e1b52f5f9ef9dc952fae994174f4ec