AW: enable SiteBlock feature by default
This flips the flag for WebViewSupervisedUserSiteBlock. Metrics for the
change are mostly neutral to positive and this change should be ready to
launch.

This does not clean up the feature flag. Code cleanup will happen later
after a full launch (as per best practices) and the old code path will
continue to be supported until that point in time.

Bug: 362574154
Test: N/A
Change-Id: I352b6d4feef82ed4f8376975f82db432d011f377
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6099112
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Peter Pakkenberg <pbirk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1397381}
NOKEYCHECK=True
GitOrigin-RevId: 88b39102d586093e73ff0d329ead54334da940bc