Specify skylab details in a nested dict in //testing/buildbot.
In //testing/buildbot/generate_buildbot_json.py, the skylab details are
read from the top-level of the test object (except within variants where
they appear in a skylab dict). This updates the generator to always read
the skylab-specific details from the skylab dict so that the treatment
of skylab is consistent in all locations and consistent with how
swarming is treated. The args argument from the skylab struct was
removed in favor of the top-level args value since args are not specific
to skylab.

The generated json still contains the skylab-specific fields at the
top-level because a recipe change would be required to update that.

Bug: 40258588
Change-Id: Ie79a0b451d19368dc1a05b5e3fb24e56344f2ff5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6159424
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1404353}
NOKEYCHECK=True
GitOrigin-RevId: 050fc4c326a2a3b3c70a6123475ca713048685e9