Make clear that headless_shell_wpt_tests is the default for WPTs
As WPT migration is mostly done and headless shell is now the default
target for WPTs, make this clear in document and naming to make it
less likely a new WPT accidentally added to the exception list.

Renamed headless_shell.filter to content_shell.filter to make it
obvious that tests listed there are intended to be run with content
shell and is something unusual.

Validate-Test-Flakiness: skip

Bug: 40927590
Change-Id: I787f7c4757aba3f869c4edeb7f43c5aeff79c41b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6154686
Reviewed-by: Rick Byers <rbyers@chromium.org>
Commit-Queue: Weizhong Xia <weizhong@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1404578}
NOKEYCHECK=True
GitOrigin-RevId: 895774587d0fa40ddfcccded9f694ae5f459ecd3