[blink scheduler] Enable DeferRendererTasksAfterInput kAllTypes policy
INP improved on all platforms with this experiment [1]. This CL updates
the default policy to kAllTypes, enables the feature, and removes the
fieldtrial config entry.

There was an increase in 99th percentile Windows keyboard responsiveness
(Blink.Responsiveness.UserInteraction.MaxEventDuration.Keyboard), but it
did not affect INP. We'll monitor the rollout and investigate the
outliers with a slow reports trigger.

[1]https://uma.googleplex.com/p/chrome/variations?sid=53556c060087209c2ba8c1fe68448342

Bug: 350540984
Change-Id: I674ac69ba8c8a8c7eb85090fbee883db758c9658
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6175333
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1406804}
NOKEYCHECK=True
GitOrigin-RevId: d601e31f08cd866ba8f661e9908a87aa76a16476