Introduce a new flag to guard adding ForcedLlamaTriggerInfo to CSPP
We want to introduce separate flags for when the ClientPhishingRequest
ping from client side detection code can contain the
LlamaForcedTriggerInfo that is from RTLookupResponse and another for
when we will use this information to trigger the on-device LLM for more
information. In addition, the flag responsible for gating the on-device
model LLM call will be added as a condition to fetch the on-device model
when the CSD service starts.

Bug: 390394312
Change-Id: I3bd58507549ffb1428630ba249451a293e05657b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6180001
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Seung Jae Lim <andysjlim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1407645}
NOKEYCHECK=True
GitOrigin-RevId: 6bc9204ea65948f955a0988dc280fb422696d428