Reland: Inherit service worker controller for srcdoc iframe
This is a reland of http://crrev.com/c/6054691 with changes from
http://crrev.com/c/6081248 for fixing the DCHECK/DwC from the original
change.

Same origin srcdoc iframes should be controlled by the same service
worker that controls the parent frame.

To achieve this, expanded ServiceWorkerClient::InheritControllerFrom to
also support srcdoc iframes and used it to setup service worker client
for srcdoc frames when NavigationRequests for them start.

To allow these clients to show up in clients.matchAll, we update same
origin check in GetWindowClients to use GetLastCommittedOrigin instead
of GetLastCommittedURL from the frames.

As the url for srcdoc iframe is about:srcdoc, when we use origin of that
url for access check, it will fail and hit DCHECK/DwC. So, we switch to
use its creator url, which is captured as url for scope match, when
performing access checks.

As the srcdoc frame test in about-blank-replacement.https.html now
passes, updated expectation to reflect that.

Also added wpt tests to ensure that service worker features work as
expected in srcdoc iframes.

The new behavior is behind the ServiceWorkerSrcdocSupport feature which
is enabled by default.

Bug: 41411856,382895651,382895896
Change-Id: I5d7e96b4838c666842315fce6795a9221e19f61b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6085871
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Liang Zhao <lzhao@microsoft.com>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1408026}
NOKEYCHECK=True
GitOrigin-RevId: 5b299cef9f6154e2a531f46f20aa2473c5f39fd1