Add fieldtrial_testing_config for EnableASWebAuthenticationSession
This CL adds the EnableASWebAuthenticationSession as we are starting its
rollout to pre-stable channel. However it is a no-op for the enabled
tests.

Per internal documentation, we cannot use ASWebAuthenticationSession for
automated tests (iOS does not allow us to interact with the web view
presented by ASWebAuthenticationSession).

The only test suite that adds accounts to the device is
ExternalURLSigninTestCase and it seems that its are disabled as they are
flaky. This means that enabling EnableASWebAuthenticationSession feature
flag has no effect for tests.

We are exploring 2 ways for re-enabling the end-to-end testing (test
suite in ExternalURLSigninTestCase):
* Continue using WKWebView temporarily, though this proves to be flaky  (all tests that add accounts to device are disabled)
* Inject the test accounts in iSL and bypass the web sign-in.
This project is ongoing, but it will also not test the add account UIs.

Bug: 373441181

Change-Id: Ib6158b4eadd865172befcf3def7b5e4db97e5509
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6179800
Reviewed-by: Arthur Milchior <arthurmilchior@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1408664}
NOKEYCHECK=True
GitOrigin-RevId: 9ccfc9dd319b556531fde4d31b9a7f13ab943a64