fuzzing: better domatolpm generation
This patch fixes the wrong creator name in DomatoLPM that led to having
incorrect variable names when storing vars in Domato context.

Also, this patch switches to using an indefinite number of lines as root
so that the mutator is responsible for correctly adjusting the number of
desired lines. This also sets an HTTPs page to the
webidl_in_process_fuzzer so that it enables APIs requiring
SecureContext.

Change-Id: I856594e815e54b1f2487201ec559a55001ef9919
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6181821
Reviewed-by: Ali Hijazi <ahijazi@chromium.org>
Commit-Queue: Ali Hijazi <ahijazi@chromium.org>
Auto-Submit: Paul Semel <paulsemel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1408711}
NOKEYCHECK=True
GitOrigin-RevId: 78cda6c6185cee8d4abd20e6e27f1c4bbd12328a