Compose GPU integration tests from mixins
Use common mixins to define the args that GPU integration tests use.
This avoids duplication of the arguments to make it easier to keep tests
in sync.

This CL starts with a group of tests that were duplicated for
ganesh vs graphite split. It also only addresses command line arguments
in mixins and not other attributes like sharding for simplicity.

Bug: 391899126
Change-Id: I03426f65e15fe15435e33c5bdfc7fc6d0ab86925
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6183105
Commit-Queue: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1410891}
NOKEYCHECK=True
GitOrigin-RevId: 67ae0c4285011bd0a8fc69e979f5c0084b45d405