Add a content scrim view
BrowserView::contents_scrim_view() is a scrim that covers the content
area. When visible, it darkens the content. This will be used as the
visual cue when showing tab-modal dialogs.

This scrim does not cover the devtools or the side panel.

screenshots,
- https://screenshot.googleplex.com/8TTc9QKUABpSLST
- https://screenshot.googleplex.com/9ZH2GNWSK3X5E53 (devtools open)
- https://screenshot.googleplex.com/3JqfcC9Z9jbhx27 (side panel open)

This CL does not include the toggling of scrim on showing tab modals.
They will come in follow up CLs.

Bug: 391903216
Change-Id: Ida3d5fa540c330a8536d1a9d279181a2276193b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6190853
Reviewed-by: Caroline Rising <corising@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Keren Zhu <kerenzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1411075}
NOKEYCHECK=True
GitOrigin-RevId: 9edaed6454cfcda778d760221c6d874e1bf9eb46