Use `run_performance_tests.py` to run Crossbench smoke tests in CQ/CI.
With a slight modification to the run_performance_tests.py tool, we can
run the Crossbench in the CQ/CI environment as well. There are two main
differences: the path to Chrome and its driver, and where the test
output is written.

Bug: 388095473
Change-Id: I54437b493d1f844457dedee06d6b3e4c761f19a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6230229
Commit-Queue: Farid (Mojtaba) Faridzad <faridzad@google.com>
Reviewed-by: John Chen <johnchen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1415935}
NOKEYCHECK=True
GitOrigin-RevId: 91af1f50db8659c2b5c353d9d9a2f6d38d7df71e