Add a feature flag to run `FontDataManager` on all pages
The webui-only experiment ran on canary/dev and is ramping up to beta
with stable coming up after.

The metrics indicate load time improvements on webui, even with the
lower statistical power of lower channels. Because of this, we want to
finish running the `FontDataService` experiment on 133 stable and
experiment with the new `FontDataManager` on all pages in 134. The goal
of this experiment is to evaluate the potential in-the-field gains of
this new font management setup.

This CL removes the webui-only feature flag and adds a new "all web
contents" flag since running the 2 in parallel would have been invalid
anyway (they each must control the typeface type chosen by the
`FontDataManager`).

Bug: 395924057
Change-Id: Iabbeb965e7736425ac1fb9580deb269646269f68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6216968
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Etienne Bergeron <etienneb@chromium.org>
Commit-Queue: Anthony Vallée-Dubois <anthonyvd@chromium.org>
Reviewed-by: David Yeung <dayeung@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1418884}
NOKEYCHECK=True
GitOrigin-RevId: 776a1489b267ba99e6548be228bede0ebc2f9a1f