Implement experimental database pre-open preloading
This change implements an experiment to preload the database prior to
opening any file handles. This resolves a conflict with Windows
exclusive file mode, which was being blocked by base::PreRead's separate
handle used for preloading.

All Databases for which the owner calls Database::Preload() are included
in the study.

This CL implements an experimental feature to preload the database
before any handles are opened.

Bug: 40904059
Change-Id: Ice9dcf730f71f5e4a6af6fd2290b9588a7e0a03a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6269958
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Austin Sullivan <asully@chromium.org>
Reviewed-by: Chris Fredrickson <cfredric@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Robert Ogden <robertogden@chromium.org>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422031}
NOKEYCHECK=True
GitOrigin-RevId: b3fd806db7d5a77681747631eace9916681b0515