[BCIV] Cleanup past experiment flags
AndroidBcivWithSuppression was a flag where we were checking if
capturing the toolbar more frequently would regress performance, it
didn't. With toolbar suppression, alot of captures were delayed until
scroll started. With BCIV, that's too late, so we had to capture the
toolbar earlier and more frequently.

AndroidBcivWithSimpleScheduler was a flag where we tried using simple
scheduler instead, it resulted in regressions.

This cl cleans up the two features. The resulting code runs as if both
are disabled.

Bug: 397711885
Change-Id: Ifbb1a74c8c7732622f97e5bc06eb65288d665911
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6281072
Commit-Queue: Peilin Wang <peilinwang@google.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1422571}
NOKEYCHECK=True
GitOrigin-RevId: 7d4920454eaadc31eb79887c80ca88d88cb14a29