[User Education] Roll User Education 2.5 out to tests.
This is a prerequisite for rolling to Canary.

This changes the field trial testing config to enable 2.5. It also adds
the ability to explicit switch between 2.0 and 2.5 in
InteractiveFeaturePromoTest.

Updates and fixes have been made to tests which assumed the 2.0
behavior.

Some systems which expected a synchronous Feature Promo Controller
behavior (even after the API was updated to be asynchronous) were also
fixed.

Bug: 369887098, 369403281
Change-Id: I04bb25486428b1b88445dfcf6d2faf257e21bb67
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6258918
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1423260}
NOKEYCHECK=True
GitOrigin-RevId: 473786be086a84ff60013720b2ec50ce7d11e248