chrome_wgsl_fuzzer: experiment with a FuzzTest grammar
This change introduces a FuzzTest grammar domain to generate better
programs for WGSL. This aims at being an experiment with grammars in
general. The introduced grammar is not complete and probably does not
work to actually parse wgsl programs. Specifically, the template
disambiguation is not correctly implemented, which is not a problem for
generating code, but is a problem for parsing.

Change-Id: I2d119af189db0a4a718c6c3d80c668d994bf05bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6304840
Auto-Submit: Paul Semel <paulsemel@chromium.org>
Reviewed-by: Ali Hijazi <ahijazi@chromium.org>
Commit-Queue: Paul Semel <paulsemel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1425645}
NOKEYCHECK=True
GitOrigin-RevId: 759480a92614276f48fd86a3e5bba098bc3e896b