[blink] Enable SkipUnnecessaryRemoteFrameGeometryPropagation by default
According to a field trial, this reduces the number of IPCs going to
RendererMain by 24% on Android, to BrowserMain by 9%, and reduces power
a tiny tiny bit (still on Android). See
https://uma.googleplex.com/p/chrome/variations?sid=3eb97711e54390d19ff169ab765b905e

As there is no tradeoff (this is stricly avoiding work), ship it.

Bug: 366424215
Change-Id: I3310220c5b31f0c92fbd2828c96553d0cf17f7f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6367056
Reviewed-by: Stefan Zager <szager@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1434206}
NOKEYCHECK=True
GitOrigin-RevId: 4514d724b2391e7a90ec96f51acb345352a58cd7