Pre-create DirectReceiver transports on Windows
Windows sandbox restrictions cause DirectReceiver's ThreadLocalNode to
crash in Transport::CreatePair. As a workaround create a transport for
the thread that's expected to use DirectReceiver before the sandbox
is locked down (which is still in single threaded context), and pass
it to ThreadLocalNode when it's assigned to the thread.

Bug: 40266729
Change-Id: Ic881eac9cf570d023a362b2cfc12f8234d706eb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6332336
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#1434954}
NOKEYCHECK=True
GitOrigin-RevId: 890090bdfc0d21035986a42d27d1438e479afa7e